+++ src/12_math.hpp
@@ -1,6 +1,7 @@
 // @begin math
-// @exports pow_mod is_prime miller_rabin sieve_spf sieve prime_list factorize divisors extgcd
-// combination crt floor_sum euler_totient euler_totient_table floor_div ceil_div mul_mod pollard 数学関連の関数群
+// @exports pow_mod is_prime miller_rabin sieve_spf sieve prime_list factorize
+// divisors extgcd combination crt floor_sum euler_totient euler_totient_table
+// floor_div ceil_div mul_mod pollard 数学関連の関数群
 namespace math {
 ll mul_mod(ll a, ll b, ll m) { return (ll)((__int128)a * b % m); }
 
@@ -43,7 +44,8 @@ bool miller_rabin(ll n) {
     }
     return true;
   };
-  for (ll a : {2LL, 325LL, 9375LL, 28178LL, 450775LL, 9780504LL, 1795265022LL}) {
+  for (ll a :
+       {2LL, 325LL, 9375LL, 28178LL, 450775LL, 9780504LL, 1795265022LL}) {
     if (a % n == 0)
       return true;
     if (witness(a))
@@ -66,7 +68,7 @@ vector<int> sieve_spf(int n) {
     for (int p : primes) {
       if (p > spf[i] || (ll)p * i > n)
         break;
-      spf[p * i] = p;
+      spf[(size_t)p * (size_t)i] = p;
     }
   }
   return spf;
